Move floder ProductDetail To Containers
